Move last updated text into left Menu component
