+++ src/components/Page/styles.tsx
@@ -81,7 +81,3 @@ export const ChapterTitleStyle = styled.h1`
   color: var(--font-color-secondary);
   font-weight: bold;
 `;
-
-export const LastUpdatedStyle = styled.p`
-  font-style: italic;
-`;
